fix: only allow the same user who originally uploaded the photo to delete it

+++ server/src/routes/places/deletePhoto.ts
@@ -1,7 +1,7 @@
 import express, { NextFunction, Request, Response } from "express";
 import { DeleteObjectCommand } from "@aws-sdk/client-s3";
 import { matchedData } from "express-validator";
-import { verifyToken } from "middleware/verifyToken";
+import { UserTokenRequest, verifyToken } from "middleware/verifyToken";
 import { getDatabase } from "db/dbSetup";
 import { respondWith, SUCCESS_MESSAGE } from "utils/responseHandling";
 import { awsS3Client } from "aws";
@@ -23,8 +23,32 @@ router.delete(
   async (req: Request, res: Response, next: NextFunction) => {
     const { fileKey } = matchedData(req) as Body;
 
+    const userId = (req as UserTokenRequest).userToken!.userId;
+
     try {
-      // First delete the photo from the DB - if this fails, then at least
+      // First make sure this user is allowed to delete the photo and that it even exists
+      const { rows: dbPhotos } = await db.query<DatabasePhoto>(
+        "SELECT * FROM photos WHERE photo_file_key = $1",
+        [fileKey]
+      );
+
+      if (dbPhotos.length === 0) {
+        return respondWith({
+          res,
+          status: 404,
+          errorMessage: "Photo does not exist",
+        });
+      }
+
+      if (dbPhotos[0].user_id !== userId) {
+        return respondWith({
+          res,
+          status: 401,
+          errorMessage: "Only photos uploaded by the same user can be deleted",
+        });
+      }
+
+      // Then delete the photo from the DB - if this fails, then at least
       // we still have the photo in S3 and don't need to roll back anything
       await db.query<DatabasePhoto>(
         "DELETE FROM photos WHERE photo_file_key = $1",
